comments and code organization

+++ ParrotSourPreProcessor.py
@@ -18,11 +18,20 @@ from alive_progress import alive_bar
 from PIL import Image
 from tqdm import tqdm
 
-OUT_DIR = "output"
-IMAGE_DIR = OUT_DIR+"\\images"
+from PSUtils import ANSWER_FILE, IMAGE_DIR, OUT_DIR, get_label, load_data
 
 
-def write_file(start_positions, n):
+def write_img(start_positions, n):
+    """
+    Plot and output a start_position image for the Nth picture
+    Parameters
+    ----------
+    start_positions : dict[]
+        An array of dict {x,y} that contains all start positions for every group
+        in every picture
+    n : int
+        The picture for which to generate an image
+    """
 
     fig, ax = plt.subplots()
     ax.scatter(start_positions[n]["x"],
@@ -50,22 +59,25 @@ if __name__ == '__main__':
 
     starttime = time.time()
 
-    shutil.rmtree(OUT_DIR)
+    # Clean and prepare output directories
+    if os.path.exists(OUT_DIR):
+        shutil.rmtree(OUT_DIR)
+    if not os.path.exists(OUT_DIR):
+        os.makedirs(OUT_DIR)
+    if not os.path.exists(IMAGE_DIR):
+        os.makedirs(IMAGE_DIR)
 
     """ IMPORT DATA """
-    data = open("trainingdata\data1000.json")
-    loaded_data = json.load(data)
+    loaded_data = load_data()
 
     """Extract numerical information"""
     Y = np.transpose(np.array([]))
 
     startPositions = {}
 
-    labels = ['AZIMUTH', 'RANGE', 'WALL',
-              "LADDER", "CHAMPAGNE", "VIC", "SINGLE"]
     n = 0
 
-    print("Generating answer key...")
+    print("Generating answer key and dataset...")
     for k in loaded_data:  # pictures
         groups = k.get("groups")
         startPositions[n] = {}
@@ -80,41 +92,28 @@ if __name__ == '__main__':
             startPositions[n]["y"] = newy
 
         # Make the answer key
-        if "pic" in k:
-            val = k.get("pic")
-
-            found_label = [label for label in labels if (label in val)][0]
-            Y = np.append(Y, found_label)
+        Y = np.append(Y, get_label(k))
         n += 1
 
-    """
-    Take X1 and X2 ---> matrix of points
-    """
-    """
-    I want one plot for every set of points in the same group {n}
-    """
-    """ plot for ML """
+    # Save the answer key
+    open(ANSWER_FILE, "w+")
+    np.savetxt(ANSWER_FILE, Y, fmt="%s")
 
     # Create scatter plots for each n_group
-    file_path = os.path.join(OUT_DIR, "Y.txt")
-
-    if not os.path.exists(OUT_DIR):
-        os.makedirs(OUT_DIR)
-    if not os.path.exists(IMAGE_DIR):
-        os.makedirs(IMAGE_DIR)
-
-    open(file_path, "w+")
-    np.savetxt(file_path, Y, fmt="%s")
-
     mpl.use("Agg")
 
-    pool = mp.Pool(mp.cpu_count()-1)
-    results = [pool.apply_async(write_file, args=([startPositions, n]))
+    # Create a pool of threads, up to the # of cores avaialble,
+    # to write image files
+    pool = mp.Pool(mp.cpu_count())
+
+    # Create images for each data row in the dataset
+    results = [pool.apply_async(write_img, args=([startPositions, n]))
                for n in range(len(startPositions))]
     pool.close()
-    output = []
+
+    # Required to track progress and get results
     for job in tqdm(results):
-        output.append(job.get())
+        job.get()
 
     total_time = time.time() - starttime
     print("Total time: " + str(total_time))
